Add outcome() function to ComputerPlayer class
This dynamic programming routine will serve as a helper for takeTurn().
It recursively checks whether the player can force a win or a draw on a
given board, storing results as they are computed to avoid duplicating work.

A player wins if he can move to a board on which he wins and draws if he cannot
win but can move to a board on which he draws.  Otherwise he loses.  The base cases
are the end game scenarios.